+++ CHANGELOG.md
@@ -7,11 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [2.5.0] - 2022-05-18
+
 ### Added
 - Assign multiple values to one property via TypoScript
 
 ### Fixed
-- Usage of stdWrap in combination with a string value in TypoScript configuration
+- Usage of stdWrap in combination with a string property value in TypoScript configuration
 
 ## [2.4.0] - 2022-03-28
 
@@ -328,7 +330,8 @@ Initial release
 - View helpers for usage in Fluid templates
 
 
-[Unreleased]: https://github.com/brotkrueml/schema/compare/v2.4.0...HEAD
+[Unreleased]: https://github.com/brotkrueml/schema/compare/v2.5.0...HEAD
+[2.5.0]: https://github.com/brotkrueml/schema/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/brotkrueml/schema/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/brotkrueml/schema/compare/v2.2.2...v2.3.0
 [2.2.2]: https://github.com/brotkrueml/schema/compare/v2.2.1...v2.2.2